solution of the Book Shop problem, rename two solutions of the Nearest Smaller Values problem and the Sliding Median problem

+++ Dynamic Programming/CoinCombinationsI.cpp
@@ -3,13 +3,13 @@
 using namespace std;
 
 int main() {
-    int n, x;
+    long n, x;
     cin >> n >> x;
-    vector<int> coins(n);
-    for (int i = 0; i < n; i++) {
+    vector<long> coins(n);
+    for (long i = 0; i < n; i++) {
         cin >> coins[i];
     }
-    vector<int> dp(x + 1, 0); // dp[i] = number of ways to make sum i
+    vector<long> dp(x + 1, 0); // dp[i] = number of ways to make sum i
     dp[0] = 1;
     /* 
     dp[i] = sum(dp[i - coins[j]]) for all j (ignore if i - coins[j] < 0). dp[i] = 0 means that there is no way to 
@@ -17,8 +17,8 @@ int main() {
     and add dp[i - coins[j]] to dp[i]. Note that i - coins[j] < i, so dp[i - coins[j]] has already been calculated
     before we calculate dp[i]. 
     */
-    for (int i = 1; i <= x; i++) {
-        for (int j = 0; j < n; j++) {
+    for (long i = 1; i <= x; i++) {
+        for (long j = 0; j < n; j++) {
             if (i - coins[j] >= 0) {
                 dp[i] += dp[i - coins[j]];
                 dp[i] %= long(1e9 + 7);
